Add files via upload
With these files, MARTY now will output a file with loop integral expressions/strings for 1-to-1 1-loop processes. 
For 2-to-2 processes, the program encounters segmentation faults for certain diagrams, and hangs up indefinitely. This may be due to how the loop integral expressions are handled as they are processed. Still very much a work-in-process.