Add files via upload
These files should process and export a file containing loop integral expressions for 1-to-1 1-loop amplitudes (in addition to amplitudes, diagrams, squared amplitudes, etc.).
The program encounters a segfault for some processes when attempting to run for 2-to-2 diagrams. This may be due to how the loop diagrams are processed. Fixing this is a work in progress.